refactor(web): RelationshipNetwork Code Splitting
